+++ History.md
@@ -21,6 +21,9 @@
 
 * `Template.parentData`'s argument defaults to 1. #2861
 
+* `Blaze.remove` on a template's view now correctly removes the DOM
+  when the template was inserted using `Blaze.renderWithData`. #3130
+
 * Upgraded dependencies:
   - node: 0.10.33 (from 0.10.29)
   - source-map-support: 0.2.8 (from 0.2.5)